Add settings view
Moved save location path there
