+++ app/components/placeholder.tsx
@@ -1,0 +1,6 @@
+export default function Placeholder() {
+  return (
+    <div className="relative flex h-full min-h-[50vh] flex-col rounded-xl bg-muted/50 p-4 lg:col-span-2">
+    </div>
+  )
+}
