+++ app/page.tsx
@@ -1,14 +1,21 @@
+'use client';
+
 import { ToastContainer } from 'react-toastify';
 
 import UI from "@/app/components/ui";
+import { ViewProvider } from '@/app/contexts/view';
 
 export default function Home() {
   return (
     <>
-      <ToastContainer
-        newestOnTop
-      />
-      <UI />
+      <ViewProvider>
+        <>
+          <ToastContainer
+            newestOnTop
+          />
+          <UI />
+        </>
+      </ViewProvider>
     </>
-  );
+  )
 }
