--- app/types/form-input.ts
@@ -1,5 +1,0 @@
-export type FormInput = {
-  url: string
-  save_to?: string
-  audio_only?: boolean
-}
