+++ app/types/form.ts
@@ -1,0 +1,8 @@
+export type MainForm = {
+  url: string
+  audio_only?: boolean
+}
+
+export type SettingsForm = {
+  save_to?: string
+}
